requirements.txt
annotated-types~=0.7.0
anyio~=4.9.0
attrs~=24.3.0
beautifulsoup4~=4.13.4
boto3~=1.37.10
botocore~=1.37.10
certifi~=2025.7.14
cffi~=1.17.1
charset-normalizer~=3.3.2
click~=8.1
cryptography>=44.0.1,<45
curl-cffi~=0.12.0
distro~=1.9.0
fastapi~=0.116.1
frozendict~=2.4.2
h11~=0.16.0
httpx~=0.28.1
idna~=3.7
jmespath~=1.0.1
lxml~=5.3.0
mangum~=0.19.0
multitasking~=0.0.12
numexpr~=2.11.0
numpy~=2.3.1
openpyxl~=3.1.5
outcome~=1.3.0
packaging~=24.1
pandas~=2.3.1
peewee~=3.18.2
pillow~=11.3.0
platformdirs~=3.10.0
protobuf~=6.31.1
pyarrow~=21.0.0
pycparser~=2.21
pyjwt~=2.9.0
pydantic~=2.11.7
pysocks~=1.7.1
python-dateutil~=2.9.0post0
pytz~=2025.2
pyyaml~=6.0.2
requests~=2.32.4
s3transfer~=0.11.2
selenium~=4.24.0
six~=1.17.0
slowapi~=0.1.9
sniffio~=1.3.1
sortedcontainers~=2.4.0
soupsieve~=2.7
starlette~=0.47.2
trio-websocket~=0.12.2
trio~=0.30.0
typing_extensions~=4.12.2
tzdata~=2025.2
urllib3~=2.5.0
uvicorn~=0.35.0
websocket-client~=1.8.0
websockets~=15.0.1
win_inet_pton~=1.1.0
wsproto~=1.2.0
yfinance~=0.2.65
google-auth~=2.36.0

google-auth-oauthlib~=1.2.2
google-auth-httplib2~=0.2.0
pytest-asyncio~=0.23
